+++ src/firestore/inject-collection.decorator.ts
@@ -1,4 +1,4 @@
-import { Inject } from '@nestjs/common';
+import { Inject, Type } from '@nestjs/common';
 
 /**
  * Returns a unique key that references a Firestore collection that should be injected.
@@ -6,9 +6,7 @@ import { Inject } from '@nestjs/common';
  * @param documentType The type of the document stored in the collection.
  * @returns The key.
  */
-export function getFirestoreCollectionInjectionName(documentType: {
-  new (): any;
-}) {
+export function getFirestoreCollectionInjectionName(documentType: Type) {
   return `CAUSA_FIRESTORE_COLLECTION#${documentType.name}`;
 }
 
@@ -17,5 +15,5 @@ export function getFirestoreCollectionInjectionName(documentType: {
  *
  * @param documentType The type of the document stored in the collection.
  */
-export const InjectFirestoreCollection = (documentType: { new (): any }) =>
+export const InjectFirestoreCollection = (documentType: Type) =>
   Inject(getFirestoreCollectionInjectionName(documentType));